Enroll command rework - repository
