+++ infra/repository.py
@@ -5,8 +5,51 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.db import transaction
 
 
+class RepositoryCrud:
+    root_model = None
+    root_entity = None
+
+    def __init__(self, user=None) -> None:
+        self.user = user
+
+    def retrieve(self, uuid: UUID):
+        try:
+            obj = self.root_model.objects.get(uuid=uuid)
+            return self._from_object(obj)
+        except self.root_model.DoesNotExist:
+            return None
+
+    def search(self, **kwargs):
+        try:
+            obj = self.root_model.objects.get(**kwargs)
+            return self._from_object(obj)
+        except self.root_model.DoesNotExist:
+            return None
+
+    def list(self):
+        return [self._from_object(obj) for obj in self.root_model.objects.all()]
+
+    def create(self, **kwargs):
+        obj = self.root_model.objects.create(**kwargs)
+        return self._from_object(obj)
+
+    def update(self, obj_uuid, **kwargs):
+        obj = self.root_model.objects.get(uuid=obj_uuid)
+        for key, value in kwargs.items():
+            setattr(obj, key, value)
+        obj.save()
+        return self._from_object(obj)
+
+    def delete(self, uuid: UUID):
+        self.root_model.objects.get(uuid=uuid).delete()
+
+    def _from_object(self, obj):
+        raise NotImplementedError("Implement this method in child class")
+
+
 class Repository:
     root_model = None
+    crud_repo: RepositoryCrud() = None
 
     def __init__(self, user=None) -> None:
         self.user = user
@@ -50,42 +93,3 @@ class Repository:
 
             if commit:
                 self.update(self.resource)
-
-
-class RepositoryCrud:
-    root_model = None
-    root_entity = None
-
-    def retrieve(self, uuid: UUID):
-        try:
-            obj = self.root_model.objects.get(uuid=uuid)
-            return self._from_object(obj)
-        except self.root_model.DoesNotExist:
-            return None
-
-    def search(self, **kwargs):
-        try:
-            obj = self.root_model.objects.get(**kwargs)
-            return self._from_object(obj)
-        except self.root_model.DoesNotExist:
-            return None
-
-    def list(self):
-        return [self._from_object(obj) for obj in self.root_model.objects.all()]
-
-    def create(self, **kwargs):
-        obj = self.root_model.objects.create(**kwargs)
-        return self._from_object(obj)
-
-    def update(self, obj_uuid, **kwargs):
-        obj = self.root_model.objects.get(uuid=obj_uuid)
-        for key, value in kwargs.items():
-            setattr(obj, key, value)
-        obj.save()
-        return self._from_object(obj)
-
-    def delete(self, uuid: UUID):
-        self.root_model.objects.get(uuid=uuid).delete()
-
-    def _from_object(self, obj):
-        raise NotImplementedError("Implement this method in child class")
